Refactor stdlib into separate package with shared object types
- Create pkg/object/ for shared types used by interpreter and stdlib
- Create pkg/stdlib/ with modular standard library implementation:
  - stdlib.go: Entry point with GetAllBuiltins() aggregator
  - builtins.go: Core std module (println, print, len, typeof, etc.)
  - math.go: Math functions (sqrt, pow, sin, cos, abs, random, etc.)
  - arrays.go: Array utilities (push, pop, map, filter, reduce, sort, etc.)
  - strings.go: String functions (upper, lower, trim, split, join, etc.)
  - time.go: Time functions (now, format, parse, sleep, etc.)
- Update interpreter to import types from object package via aliases
- Remove old lib_*.go files from interpreter package

This separation enables cleaner architecture and easier stdlib expansion.
