+++ pkg/interpreter/builtins.go
@@ -4,15 +4,18 @@ package interpreter
 // Licensed under the MIT License. See LICENSE for details.
 
 import (
-	"bufio"
-	"fmt"
-	"os"
-	"strconv"
-	"strings"
+	"github.com/marshallburns/ez/pkg/object"
+	"github.com/marshallburns/ez/pkg/stdlib"
 )
 
-// Global stdin reader to maintain buffering across multiple input() calls
-var stdinReader = bufio.NewReader(os.Stdin)
+// builtins maps function names to their implementations.
+// The actual implementations are in pkg/stdlib/.
+var builtins map[string]*object.Builtin
+
+func init() {
+	// Get all builtins from the stdlib package
+	builtins = stdlib.GetAllBuiltins()
+}
 
 // getEZTypeName returns the EZ language type name for an object
 // For integers, returns the declared type (e.g., "u64", "i32") instead of generic "INTEGER"
@@ -44,214 +47,4 @@ func getEZTypeName(obj Object) string {
 	default:
 		return string(obj.Type())
 	}
-}
-
-var builtins = map[string]*Builtin{
-	// BASIC STANDARD LIBRARY BUILTINS
-
-	// Prints values to standard output followed by a newline. Any number of arguments can be provided.
-	"std.println": {
-		Fn: func(args ...Object) Object {
-			for i, arg := range args {
-				if i > 0 {
-					fmt.Print(" ")
-				}
-				fmt.Print(arg.Inspect())
-			}
-			fmt.Println()
-			return NIL
-		},
-	},
-	// Prints values to standard output WITHOUT a newline. Any number of arguments can be provided.
-	"std.print": {
-		Fn: func(args ...Object) Object {
-			for i, arg := range args {
-				if i > 0 {
-					fmt.Print(" ")
-				}
-				fmt.Print(arg.Inspect())
-			}
-			return NIL
-		},
-	},
-
-	// START OF FULL BUILTINS
-
-	// Reads an integer from standard input. Returns the integer and an error object (nil if no error).
-	// Recomended to just use input() and convert to type: `int` manually.
-	"read_int": {
-		Fn: func(args ...Object) Object {
-			reader := bufio.NewReader(os.Stdin)
-			text, readErr := reader.ReadString('\n')
-			// Trim newline
-			if len(text) > 0 && text[len(text)-1] == '\n' {
-				text = text[:len(text)-1]
-			}
-
-			// Handle read error
-			if readErr != nil {
-				errObj := &Struct{
-					TypeName: "Error",
-					Fields: map[string]Object{
-						"message": &String{Value: "failed to read input"},
-						"code":    &Integer{Value: 1},
-					},
-				}
-				return &ReturnValue{Values: []Object{&Integer{Value: 0}, errObj}}
-			}
-
-			// Parse integer
-			val, parseErr := strconv.ParseInt(text, 10, 64)
-			if parseErr != nil {
-				errObj := &Struct{
-					TypeName: "Error",
-					Fields: map[string]Object{
-						"message": &String{Value: "invalid integer input"},
-						"code":    &Integer{Value: 2},
-					},
-				}
-				return &ReturnValue{Values: []Object{&Integer{Value: 0}, errObj}}
-			}
-
-			// Success - return value and nil error
-			return &ReturnValue{Values: []Object{&Integer{Value: val}, NIL}}
-		},
-	},
-
-	// Returns the length of a string or array as an integer.
-	"len": {
-		Fn: func(args ...Object) Object {
-			if len(args) != 1 {
-				return &Error{Code: "E7001", Message: fmt.Sprintf("len() takes exactly 1 argument, got %d", len(args))}
-			}
-
-			switch arg := args[0].(type) {
-			case *String:
-				return &Integer{Value: int64(len(arg.Value))}
-			case *Array:
-				return &Integer{Value: int64(len(arg.Elements))}
-			default:
-				return &Error{Code: "E7003", Message: fmt.Sprintf("len() not supported for %s", args[0].Type())}
-			}
-		},
-	},
-	// Returns the type of the given object as a string.
-	// For integers, returns the declared type (e.g., "u64", "i32") instead of generic "int"
-	"typeof": {
-		Fn: func(args ...Object) Object {
-			if len(args) != 1 {
-				return &Error{Code: "E7001", Message: "typeof() takes exactly 1 argument"}
-			}
-			// Use the helper function to get the EZ type name
-			return &String{Value: getEZTypeName(args[0])}
-		},
-	},
-
-	// Converts a value to an integer if possible.
-	"int": {
-		Fn: func(args ...Object) Object {
-			if len(args) != 1 {
-				return &Error{Code: "E7001", Message: "int() takes exactly 1 argument"}
-			}
-			switch arg := args[0].(type) {
-			case *Integer:
-				return arg
-			case *Float:
-				return &Integer{Value: int64(arg.Value)}
-			case *String:
-				// Strip underscores to support underscore integer syntax (e.g., "100_000")
-				cleanedValue := strings.ReplaceAll(arg.Value, "_", "")
-				val, err := strconv.ParseInt(cleanedValue, 10, 64)
-				if err != nil {
-					return &Error{
-						Code: "E7005",
-						Message: fmt.Sprintf("cannot convert %q to int: invalid integer format\n\n"+
-							"The string must contain only digits (0-9), optionally with:\n"+
-							"  - A leading + or - sign\n"+
-							"  - Underscores for readability (e.g., \"100_000\")\n\n"+
-							"Examples of valid integers:\n"+
-							"  \"42\", \"-123\", \"1_000_000\"", arg.Value),
-					}
-				}
-				return &Integer{Value: val}
-			case *Char:
-				return &Integer{Value: int64(arg.Value)}
-			default:
-				if args[0].Type() == ARRAY_OBJ {
-					return &Error{
-						Code: "E7005",
-						Message: "cannot convert ARRAY to int\n\n" +
-							"Arrays cannot be directly converted to integers.\n" +
-							"Hint: If you want the array length, use len() instead:\n" +
-							"    len(myArray)  // returns the number of elements",
-					}
-				}
-				return &Error{Code: "E7005", Message: fmt.Sprintf("cannot convert %s to int", args[0].Type())}
-			}
-		},
-	},
-	// Converts a value to a float if possible.
-	"float": {
-		Fn: func(args ...Object) Object {
-			if len(args) != 1 {
-				return &Error{Code: "E7001", Message: "float() takes exactly 1 argument"}
-			}
-			switch arg := args[0].(type) {
-			case *Float:
-				return arg
-			case *Integer:
-				return &Float{Value: float64(arg.Value)}
-			case *String:
-				// Strip underscores to support underscore number syntax (e.g., "3.14_159")
-				cleanedValue := strings.ReplaceAll(arg.Value, "_", "")
-				val, err := strconv.ParseFloat(cleanedValue, 64)
-				if err != nil {
-					return &Error{
-						Code: "E7006",
-						Message: fmt.Sprintf("cannot convert %q to float: invalid float format\n\n"+
-							"The string must be a valid floating-point number with:\n"+
-							"  - Optional leading + or - sign\n"+
-							"  - Digits with optional decimal point\n"+
-							"  - Underscores for readability (e.g., \"3.14_159\")\n"+
-							"  - Optional scientific notation (e.g., \"1.5e10\")\n\n"+
-							"Examples of valid floats:\n"+
-							"  \"3.14\", \"-2.5\", \"1_000.50\", \"1.5e10\"", arg.Value),
-					}
-				}
-				return &Float{Value: val}
-			default:
-				if args[0].Type() == ARRAY_OBJ {
-					return &Error{
-						Code: "E7006",
-						Message: "cannot convert ARRAY to float\n\n" +
-							"Arrays cannot be directly converted to floating-point numbers.\n" +
-							"Hint: If you want the array length, use len() instead:\n" +
-							"    len(myArray)  // returns the number of elements",
-					}
-				}
-				return &Error{Code: "E7006", Message: fmt.Sprintf("cannot convert %s to float", args[0].Type())}
-			}
-		},
-	},
-
-	// Converts a value to a string.
-	"string": {
-		Fn: func(args ...Object) Object {
-			if len(args) != 1 {
-				return &Error{Code: "E7001", Message: "string() takes exactly 1 argument"}
-			}
-			return &String{Value: args[0].Inspect()}
-		},
-	},
-
-	// Reads a line of input from standard input and returns it as a string.
-	"input": {
-		Fn: func(args ...Object) Object {
-			// Use global reader to avoid buffering issues with multiple input() calls
-			text, _ := stdinReader.ReadString('\n')
-			// Trim newline characters (both \n and \r\n for cross-platform support)
-			text = strings.TrimRight(text, "\r\n")
-			return &String{Value: text}
-		},
-	},
 }
